Fixed the issue with failing to get messages which was spamming the log - it was caused by trying to get a message in a channel without the `read message history` permission.
